+++ tests/test_functional/test_retry_if_none.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 import unittest.mock
 
-import pymince.retry
+import pymince.functional
 
 
 def test_returns_none_with_tries():
     func = unittest.mock.Mock(return_value=None)
-    decorated_func = pymince.retry.retry_if_none(tries=3)(func)
+    decorated_func = pymince.functional.retry_if_none(tries=3)(func)
     res = decorated_func()
     assert func.call_count == 3
     assert res is func.return_value
@@ -14,7 +14,7 @@ def test_returns_none_with_tries():
 
 def test_returns_none_without_tries():
     func = unittest.mock.Mock(return_value=None)
-    decorated_func = pymince.retry.retry_if_none()(func)
+    decorated_func = pymince.functional.retry_if_none()(func)
     res = decorated_func()
     assert func.call_count == 1
     assert res is func.return_value
@@ -22,7 +22,7 @@ def test_returns_none_without_tries():
 
 def test_returns_not_none():
     func = unittest.mock.Mock()
-    decorated_func = pymince.retry.retry_if_none(tries=3)(func)
+    decorated_func = pymince.functional.retry_if_none(tries=3)(func)
     res = decorated_func()
     assert func.call_count == 1
     assert res is func.return_value